fixed #13 by removing dashboard app and making some additional changes
