+++ Algolyzer/Algolyzer/urls.py
@@ -15,7 +15,6 @@ Including another URLconf
     2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
 """
 
-from dashboard import views as dashboard_views
 from django.contrib import admin
 from django.urls import include, path
 
@@ -24,5 +23,4 @@ urlpatterns = [
     path("", include("home.urls")),
     # allauth urls
     path("accounts/", include("allauth.urls")),
-    path("accounts/profile/", dashboard_views.profile, name="profile"),
 ]
