--- Algolyzer/dashboard/views.py
@@ -1,7 +1,0 @@
-from django.contrib.auth.decorators import login_required
-from django.shortcuts import render
-
-
-@login_required
-def profile(request):
-    return render(request, "dashboard/profile.html")
