+++ Algolyzer/home/templates/home/dashboard.html
