+++ Algolyzer/home/templates/home/home.html
@@ -1,0 +1,7 @@
+<h1>This is HOME page view</h1>
+
+{% if not user.is_authenticated %}
+    <a href="{% url 'account_login' %}">Login</a>
+{% endif %}
+
+<a href="{% url 'dashboard' %}">Dashboard</a>
