+++ Algolyzer/home/urls.py
@@ -1,7 +1,8 @@
 from django.urls import path
 
-from .views import HomePageView
+from .views import dashboard, home
 
 urlpatterns = [
-    path("", HomePageView.as_view(), name="home"),
+    path("", home, name="home"),
+    path("dashboard/", dashboard, name="dashboard"),
 ]
