+++ Algolyzer/home/views.py
@@ -1,6 +1,13 @@
-# from django.shortcuts import render
-from django.views.generic import TemplateView
+from django.contrib.auth.decorators import login_required
+from django.shortcuts import render
 
 
-class HomePageView(TemplateView):
-    template_name = "home.html"
+def home(request):
+    """Main Home Page"""
+    return render(request, "home/home.html")
+
+
+@login_required
+def dashboard(request):
+    """Dashboard page for logged in users only"""
+    return render(request, "home/dashboard.html")
